--- Modules/Domain/Auth/Project.swift
@@ -1,17 +1,0 @@
-import ProjectDescription
-
-let project = Project(
-    name: "AuthDomain",
-    targets: [
-        .target(
-            name: "AuthDomain",
-            destinations: .iOS,
-            product: .staticFramework,
-        
-            bundleId: "com.hereit.domain.auth",
-            infoPlist: .default,
-            sources: ["Sources/**"],
-            dependencies: []
-        )
-    ]
-)
