+++ Projects/Data/Project.swift
@@ -1,0 +1,16 @@
+import ProjectDescription
+
+let project = Project(
+    name: "Data",
+    targets: [
+        .target(
+            name: "Data",
+            destinations: .iOS,
+            product: .staticFramework,
+            bundleId: "com.hereit.data",
+            infoPlist: .default,
+            sources: ["Sources/**"],
+            dependencies: []
+        )
+    ]
+)
