+++ Projects/Presentation/Resources/Assets.xcassets/Contents.json
