+++ Projects/Shared/Resources/Assets.xcassets/Contents.json
