+++ Projects/Shared/Resources/Fonts/Pretendard-Regular.ttf
